Threading
Playing with threading on elements of pixel draw - perl threads are
frankly a bit poor especially if it involves Tk. It works but; there is
currently a problem with threads being left unjoined which in turn could
become a resource issue. There is a big comment about this in the
_drawObjectPixel function